some refactoring (common methods to Utils and tests improvements)
