+++ konduit-serving-pipeline/src/main/java/ai/konduit/serving/pipeline/impl/step/ml/regression/RegressionOutputRunner.java
@@ -16,22 +16,22 @@
  *  *****************************************************************************
  */
 
-package ai.konduit.serving.pipeline.impl.step.ml.ssd.regression;
+package ai.konduit.serving.pipeline.impl.step.ml.regression;
 
 import ai.konduit.serving.annotation.runner.CanRun;
 import ai.konduit.serving.pipeline.api.context.Context;
 import ai.konduit.serving.pipeline.api.data.Data;
 import ai.konduit.serving.pipeline.api.data.NDArray;
-import ai.konduit.serving.pipeline.api.data.NDArrayType;
 import ai.konduit.serving.pipeline.api.data.ValueType;
 import ai.konduit.serving.pipeline.api.step.PipelineStep;
 import ai.konduit.serving.pipeline.api.step.PipelineStepRunner;
 import ai.konduit.serving.pipeline.util.DataUtils;
 import lombok.AllArgsConstructor;
 
-import java.math.BigDecimal;
-import java.math.RoundingMode;
 import java.util.*;
+
+import static ai.konduit.serving.pipeline.impl.step.ml.Utils.FloatNDArrayToDouble;
+import static ai.konduit.serving.pipeline.impl.step.ml.Utils.squeeze;
 
 @AllArgsConstructor
 @CanRun(RegressionOutputStep.class)
@@ -65,7 +65,7 @@ public class RegressionOutputRunner implements PipelineStepRunner {
 
         NDArray regressionOutput = data.getNDArray(inputName);
         if (regressionOutput.shape().length > 2) {
-            throw new UnsupportedOperationException("Invalid input to ClassifierOutputStep: only rank 1 or 2 inputs are available, got array with shape" + Arrays.toString(regressionOutput.shape()));
+            throw new UnsupportedOperationException("Invalid input to RegressionOutputStep: only rank 1 or 2 inputs are available, got array with shape" + Arrays.toString(regressionOutput.shape()));
 
         }
         regressionOutput = FloatNDArrayToDouble(regressionOutput);
@@ -84,7 +84,7 @@ public class RegressionOutputRunner implements PipelineStepRunner {
 
 
         if (!batch) {
-            double[] regressionOutputArr = squeze(regressionOutput);
+            double[] regressionOutputArr = squeeze(regressionOutput);
             for (Map.Entry<String, Integer> entry : outputNames.entrySet())
                 data.put(entry.getKey(), regressionOutputArr[entry.getValue()]);
 
@@ -109,46 +109,6 @@ public class RegressionOutputRunner implements PipelineStepRunner {
 
 
         return data;
-    }
-
-
-    private static double[] squeze(NDArray arr) {
-
-        // we have [numClasses] array, so do not modify nothing
-        if (arr.shape().length == 1) {
-            return arr.getAs(double[].class);
-        }
-
-        // i.e we have [1, numClasses] array
-        if (arr.shape().length == 2 && arr.shape()[0] == 1) {
-            return arr.getAs(double[][].class)[0];
-        }
-
-        return null;
-
-    }
-
-    private static NDArray FloatNDArrayToDouble(NDArray ndarr) {
-        if (ndarr.type() == NDArrayType.FLOAT || ndarr.type() == NDArrayType.FLOAT16 || ndarr.type() == NDArrayType.BFLOAT16) {
-            float[][] farr = ndarr.getAs(float[][].class);
-            double[][] darr = new double[(int) ndarr.shape()[0]][(int) ndarr.shape()[1]];
-            for (int i = 0; i < farr.length; i++) {
-                for (int j = 0; j < farr[i].length; j++) {
-                    darr[i][j] = round(Double.valueOf(farr[i][j]),7);
-                }
-            }
-
-            return NDArray.create(darr);
-        }
-        return ndarr;
-    }
-
-    private static double round(double value, int places) {
-        if (places < 0) throw new IllegalArgumentException();
-
-        BigDecimal bd = new BigDecimal(Double.toString(value));
-        bd = bd.setScale(places, RoundingMode.HALF_UP);
-        return bd.doubleValue();
     }
