+++ konduit-serving-pipeline/src/main/resources/META-INF/services/ai.konduit.serving.pipeline.api.step.PipelineStepRunnerFactory
@@ -18,5 +18,5 @@ ai.konduit.serving.pipeline.impl.step.logging.LoggingPipelineStepRunnerFactory
 ai.konduit.serving.pipeline.impl.step.ml.ssd.SSDToBoundingBoxRunnerFactory
 ai.konduit.serving.pipeline.impl.step.bbox.filter.BoundingBoxFilterStepRunnerFactory
 ai.konduit.serving.pipeline.impl.step.bbox.point.BoundingBoxToPointStepRunnerFactory
-ai.konduit.serving.pipeline.impl.step.ml.ssd.regression.RegressionOutputRunnerFactory
+ai.konduit.serving.pipeline.impl.step.ml.regression.RegressionOutputRunnerFactory
 ai.konduit.serving.pipeline.impl.step.ml.classifier.ClassifierOutputRunnerFactory
